FIX: minor variable fixes

+++ api/brand/Controller.js
@@ -52,31 +52,51 @@ const getBrandByName = async (req, res) => {
 
 ///api/brand/createBrand
 const createBrand = async (req, res) => {
-	const { Name, Image } = req.body;
-	try {
-		// Connection to database
-		await connect(process.env.MONGO_URI);
-		Brands.create({ name: Name, image: Image });
-		res.status(201).json({ message: "Success" });
-	} catch (error) {
-		res.json({
-			message: error.message,
-		});
+	const { name, image } = req.body;
+	if (!name || !image) {
+		res.status(400).json({
+			message: "Oops! Missing Required Field."
+		})
+	} else {
+		try {
+			// Connection to database
+			await connect(process.env.MONGO_URI);
+			//search brand
+			const CheckBrand = await Brands.findOne({ name })
+			//if brand found, stop
+			if (CheckBrand) {
+				res.json({
+					message: "Oops! Brand Already Exists."
+				})
+			}
+			Brands.create({ name, image });
+			res.status(201).json({ message: "Success" });
+		} catch (error) {
+			res.json({
+				message: error.message,
+			});
+		}
 	}
 };
 
 ///api/brand/deleteBrand
 const deleteBrand = async (req, res) => {
 	const { Name } = req.body;
-	try {
-		// Connection to database
-		await connect(process.env.MONGO_URI);
-		await Brands.deleteOne({ name: Name });
-		res.json({ message: "Success" });
-	} catch (error) {
-		res.json({
-			message: error.message,
-		});
+	if (!Name) {
+		res.status(400).json({
+			message: "Oops! Missing Required Field."
+		})
+	} else {
+		try {
+			// Connection to database
+			await connect(process.env.MONGO_URI);
+			await Brands.deleteOne({ name: Name });
+			res.json({ message: "Success" });
+		} catch (error) {
+			res.json({
+				message: error.message,
+			});
+		}
 	}
 };
